packages/python-sdk/pyproject.toml
[tool.poetry]
name = "hyrelog-python"
version = "1.0.0"
description = "Official HyreLog Python SDK"
authors = ["HyreLog <support@hyrelog.com>"]
readme = "README.md"
packages = [{include = "hyrelog"}]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
pydantic = "^2.5.0"
opentelemetry-api = "^1.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.12.1"
mypy = "^1.7.1"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py39"

